Use default Haxe, Neko, and Node.js install locations, if available.
This reduces the amount of manual configuration a new Moonshine user needs to do for Haxe development (and for anything else that uses Node.js). If they don't like these defaults, they can change the settings normally, which is already what they would have needed to do, so it won't hurt anything. If the user has already saved these paths in settings (or if the path has been "cleared"), nothing will be updated. In other words, it only affects new installations. My previous commit that sets a default path will automatically set the open folder dialog path to make it easier for existing users too.

This behavior is skipped in the macOS App Store version to avoid sandbox restrictions.