Language servers: Fixed encoding of .exe path used in .cmd file on Windows (references #1066)
Partially reverts commit f1f5b3b08fb036b0b89c1add95380fa8c12bb46c

It was causing an error with certain paths that also contained spaces. The error I encountered appears below:

The filename, directory name, or volume label syntax is incorrect.

What was happening was that path was getting wrapped in quotes because there was a space, but it was also being encoded with the ^ character. The ^ character is not considered an escape character if it is wrapped in quotes, so it was being treated as a part of the path.

This change removes the call to UtilsCore.getEncodedForShell(), but it expands CommandLineUtil.escapeSingleOption() to handle more characters that are reserved by cmd.exe batch files and require quotes.

I tested and confirmed that this fixes cases where the path contains reserved ( and ) characters and where it contains spaces (and where it contains both reserved characters and spaces).