Fixes #1108
It’s a general data-binding issue, which feature is non-existent in the Haxe/OpenFL environment. I’ve implemented an EventDispatcher and the relevant event listeners in this specific component, so the data change is now properly handled. But I’m going to need to find a way to implement the data-binding concept for the entire project (currently WIP)