Build: ensure that AIR_SDK is defined for OpenFL
This is not strictly needed for local development, but it's a good idea for a build script that should be as consistent as possible. Adding this argument ensures that the same Flex/AIR SDK is used for both building the SWF and for packaging the AIR bundle.

Locally, as long as the developer has run `lime config AIR_SDK <path to AIR SDK>`, then they should be able to use `lime build air -debug` without this extra argument.