Java: fix update mechanism that stopped working because the launcher JAR file name might not always change between versions
In the past, the launcher JAR file name was always updated with a new version number (even if it didn't match the eclipse.jdt.ls version number), but it looks like that isn't guaranteed. However, there's another org.eclipse.jdt.ls.core JAR file that definitely will be updated with every release, so we're switching to that one instead.