MainEntryPoint: call haxe.initSwc(null)
This is required by Haxe SWC files. We seemed to be able to get away without calling it for a long time, but I found at least one exception thrown with Haxe 4.3 that is fixed by adding this call. It's possible that this will fix other unexplained exceptions that don't happen often enough to have been worth reporting as bugs.